+++ apps/dw-frontend/requirements.txt
@@ -1,4 +1,4 @@
-streamlit>=1.46.1
+streamlit>=1.47.0
 plotly>=5.18.0
 pandas>=2.2.0
 python-dateutil>=2.8.2
