+++ apps/dw-frontend/utils/constants.py
@@ -1,0 +1,2 @@
+# Shared constants
+API_BASE = "http://localhost:4200/consumption"
